Every script calls Monocle.pieceLoaded.
You can use this to detect when parts of Monocle have loaded, in a
similar way to how JSON-P works. If you define the global function
onMonoclePiece, it will be passed the piece name.

As a result of this, the framer functionality now works in IE8 (which
bizarrely evaluates script tags in dynamic iframes when earlier script
tags have not been loaded).