Refined reader initialisation and book turn-to-first-page flow.
Reader now takes a place option (which should be a valid locus object).
PlaceSaver has been modified so that it can be initialised before the
reader and generate this locus object for the place option.

Page changes on a pageDiv while that pageDiv is already changing are now
explicitly rejected. May add a queuing system for pageChange requests,
if there's a use case for it?