The iOS4 page-turn flicker is vanquished at last.
The problem stems from applying a transform with a non-immediate
transition when there is already a transform with a transition occuring.
This used to be fine, now it's not. Answer was simply to zero the
duration of the transition when the page position is tracking the
position of the cursor.