Full rewrite of the page turning logic. WIP.
The page turning logic affects:

  Reader (setPage)
  Book (calculating the page to turn to, handling cmpt loading)
  Component (actually loading content into pageDiv frames)
  Flippers (updating the offset of the frame to curr page)

Trying to rationalise all of this stuff. Removing Reader class
from the process. Reducing the complexity of the communication
between the Book and the Component. Shifting the calculation of
the pixel offset into the Flipper (rather than the Book).

At the moment, Slider flipper is busted, and of course the Legacy
flipper is still busted.

The pagechange/pagechanging events are currently disabled.