Working around the iOS select-thru-iframe bug.
This was causing touch events to hit the second iframe in the DOM,
irrespective of which iframe was actually on top (in z-index order). So
"behind-text" was getting selected.

Now, the second iframe is always the top iframe, except during
animations. So touch events should always hit the correct text in
interactive mode.