A few more initialisation-time events.
The two major periods of waiting during initialisation are while the
iframe in each page is "primed" to a particular URL, and while it is set
to its first component.

So, eg, if you use the slider flipper, you might have to wait for four
component loads before Reader is initialised: the priming and setting of
both frames.

Now you can listen for those, with 'monocle:frameprimed' and
'monocle:firstcomponentchange'.

Example uses are progress bars, or benchmarking the speed of
initialisation.