Stencil default click behaviour affects links in interactive mode too.
This was temporarily disabled so that any click listeners on the
original link could perform the handling. But of course then you don't
get nice skipToChapter overrides.

So now the links in the original documents have a stencilClickHandler
property added to them, which you can deafen if you don't want the
default stencil click behaviour in interactive mode.