Make the default style rules harder to accidentally override.
Encountered a style: .bodycenter { margin: 25% !important; }, that broke
the body element because it overrode the default style body { margin: 0
!important; }. By making our rule more specific, it's less likely that
this sort of thing will happen.

Of course, it remains an arms race. Applying inline styles to every
single element is the nuclear option -- there's no intention of doing
that though.