A workaround for the Chrome 10 blank pages bug.
Discussion of bug is here:

   https://github.com/joseph/monocle/issues#issue/25

The bug affects WebKit nightlies since December 2010 and Chrome since
version 10. Reasonable expectation of fix in Chrome 11, and that it
won't ever appear in Safari.

The workaround is problematic because:

* it significantly modifies the DOM of the component
* it hides all scrollbars within the component

But it's a lot better than blank pages. Thanks to madfarmer for
identifying the workaround.