JavaScript load strategy for components.
If you return content from your bookdata getComponent like this:

    { javascript: "... a string ..." }

... it will be loaded much like if you had just returned a string or
a html result. The difference is that regexes won't be run over the
string to make it 'safe' in a JavaScript statement before it is loaded
into the component frame each time. This will speed things up if you are
storing components locally — clean them once before storing them, so
that Monocle doesn't have to clean them up every time.