Recover reasonably gracefully if component fails to load.
A component is considered to have failed to load if getComponent returns
false (or the callback is called with false as the argument).

Monocle will dispatch a "monocle:componentfailed" event and try to load
the last good component instead. Listen for "monocle:componentfailed" on
the reader to alert the user to the problem.