Caching component width measurements.
This is a reasonably significant change -- the Columns dimension will no
longer automatically detect if a component is longer or shorter due to
window resize or stylesheet changes or new DOM content or whatever.

Now you need to call Reader#recalculateDimensions() explicitly whenever
you do this. This is automatically done when calling resized or changing
stylesheets using the methods provided in Reader. For other changes that
may affect component size, call this method after.