Removing support for iOS < 4.2.
This gets rid of the TouchMonitor workaround, of which I'm very proud,
but which is unquestionably bloat. iOS 4.2 is the last version to
support the iPhone 3G, which is definitely an intended target. It's too
difficult to support the original iPhone and 1st/2nd-gen iPod Touches
going forward — use the last version of the Monocle 2.0 series if you
want to support iOS 3.1 -> 5.0.