Apply all CSS modifications to body at once.
This decouples component measurement from page turns. Components are now
measured on change and whenever Reader#recalculateDimensions is
explicitly invoked.

Still a bit of a work in progress, as there are some problematic race
conditions on slower devices.

This change also fixes #22, making componentchange fire when all style
changes and measurements have been made to the component (at least,
that's the theory).