eInk Panel (supports Kindle3).
If you provide a flipper option and/or a panel option to Monocle.Reader,
these ALWAYS apply -- the legacy flipper no longer overrides it. So if
you do provide these options, it's up to you to inspect the browser
environment beforehand (such as by calling Monocle.Browser.survey
manually) and provide the appropriate option.

But by default, the legacy flipper is given to non-column browsers, and
the instant flipper is given to eInk browsers. The instant flipper will
by default assign the eInk panel to these browsers too.

Also, Browser.env.isCompatible(). Monocle now aborts and fires
"monocle:incompatible" if a browser fails to pass this test. You can
stop this event if you want to handle it yourself.

floatsIgnoreColumns workaround is back in effect.