API-CHANGE: controlshow/hide events now prefixed with monocle:
If you are listening to these events, you need to update your code to
listen to 'monocle:controlshow' and 'monocle:controlhide' instead.

The 'evt.m' object is no longer the control object itself -- it's now an
object with a 'control' property, whose value is the control object.

This brings these two events into conformance with other Monocle custom
events.