Force break-word to apply to component elements under iOS (fixes #70).
Huge thanks to @aronwoost for tracking down the cause of the problem,
which was that apparently the order in which the CSS is applied is
important. If an element has break-word, but the container doesn't have
a width (and an absolute position), it does not apply. It still doesn't
apply if the container later gets a width/position.

The fix was to apply all the CSS at the same time.