Adding column-fill:auto to pagination style rules.
This change is forced by Firefox 12:

  https://bugzilla.mozilla.org/show_bug.cgi?id=725485

Which has been updated:

  http://hg.mozilla.org/mozilla-central/rev/93b804e5f3f5

To follow the W3C css-multicol specification for column-fill:

  http://www.w3.org/TR/css3-multicol/#column-fill

I'm sure there's a good reason why the default value for column-fill is
'balance', not 'auto', but it eludes me. Anyway, Monocle was broken in
Firefox alpha releases -- this fixes it.