Component weights: use relative sizes for more accurate book percentages.
When getting the percentageOfBook from a place, or creating a locus from the
percentage of a book, Monocle assumes that all components are the same size.
Thus, if there are 20 components and you want the 10th one, you'll find it
at 50% of the book.

If you supply an array of floats for getMetaData('componentWeights'), these
will be treated as the relative sizes of the components. eg:

    [0.1, 0.1, 0.4, 0.15, 0.5, 0.2]

In this case, your six components will be factored into percentage
calculations differently (component 3 corresponds to 58%, eg).

How you calculate these weights is up to you. One option is to divide the
byte size of the component by the total bytes of all components in the book.