Refactoring component loading strategies, including injecting base URI.
The "nodes" and "javascript" strategies are now transformed into a "html"
strategy behind the scenes. In practice, there's no reason to ever use the
"javascript" strategy, so it's now deprecated.

The "url" strategy is the most natural strategy, and no baseURI modifications
are necessary. However, it's slower than the other strategies, and more
difficult to do content protection (if you need that).

The other two strategies, "html" and "doc", are applied to an iframe that has
been primed with "about:blank". This means it has the baseURI of its parent
document. If there are any relative hrefs to resources in the component
(scripts, stylesheets, images, etc), and the component should have a different
baseURI, these resources will ordinarily not load.

To date, Monocle has said "this is the bookData object's problem". No longer.
Now Monocle automatically injects a <base> tag into components loaded with the
"doc" or "html" strategy. You can override this by inserting a <base> tag
manually, or by listening on the reader for a monocle:component:baseuri event
and calling preventDefault() on it.