Added a test to confirm that SVG is working for components.
There was a problem in Chrome with the previous HTML/JavaScript load
strategies - I believe it was because Chrome wasn't processing the component as
XHTML, and therefore didn't like the xlink:href attribute from the SVG
namespace. This test confirms that the problem no longer occurs with Monocle's
new HTML load strategy.