Move for not all code paths hitting method
Not all code paths were hitting this (and calling turned) in the event that a TOC Link
was hit while residing on the page the link points to, it did not call the path to launch the turned() method.  Moved it up here and that fixed it for us.  I suspect the issue
depends on the fact that we load the entire book at once, as well.