Slightly refactored the relationship between percentageThrough and pages.
This has altered the existing behaviour - whereas Monocle.Place used to anchor
percentageThrough to the bottom of the page, now it is anchored (by default)
to the top of the page + 10% of the page. ie, "just a little way in" to the
page. This results in a more expected experience, especially when resizing the
window or the font-size.

If you need to preserve the old Monocle behaviour, set:

    Monocle.Place.PAGE_ANCHOR = 'bottom';