Fix pagination for Chrome 22+ when zooming
When using zoom in/out in chrome 22+ with sub-pixel rendering,
pages can be cut on the left or right.
A first try to fix it can be found here :
 https://github.com/joseph/Monocle/pull/146

This is a new version reducing DOM access by computed webkit
scale value only when value is updated (updating zoom fire a
window resize event).