Support a variance for basic onTap detection
Some devices are not tolerant of fat finger touches and will issue a move event in between a down/start and up/end event. This causes "onTap" to not perform well in some scenarios.

Supporting a tolerance for what constitutes as a "tap" provides the a fill gap for devices that issue erroneous events. This was mostly observed on the surface tablet and some instances of windows 8 touch devices.

The threshold can be adjusted by changing ```Gala.TAP_MAX_CONTACT_DISTNACE```.