add radix parameter to parseInt
While I'm not a huge fan of parseInt in most scenarios and prefer Number() over parseInt, I feel it's important to include the radix parameter because I've also run into subtle bugs of numbers being parse incorrectly due to radix being guessed at 2 or 8.

Not sure how you feel about this or how your preference stands but I feel it's more than dogma to ensure that the radix parameter is provided…but I also prefer to use Number() in most scenarios. :)

Let me know if this makes sense to pull in.