Force a pixel width on the component iframe.
This is a small change with a reasonably significant impact and a good reason.
For years we have set the iframe to 100% width, which is normally the width of
the parent element (the "sheaf") -- but iOS uses the width of the iframe
itself.

We've been able to work around this by scrolling the sheaf itself, which was
necessary for really old versions of iOS -- but is not necessary since at
least iOS 6.

Still, it was working before, so why mess with it? Simply, in iOS, without
a pixel width on the iframe, media queries inside the iframe are inaccurate.
A query like `@media screen and (min-width: 600)` would be false on the
iPhone because the actual "screen width" of the iframe was the width of all
of the pages -- potentially thousands and thousands of pixels.

This change makes iOS6 and 7 work the way the rest of the universe works, and
means that media queries on those devices work as you'd expect now.